Add Shape pointer to HitRecord and Material to Shape
